backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
georasters==0.5.22
tifffile==2021.8.30
numpy==1.21.2
Pillow==8.3.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56